requirements.txt
gigachat
python-telegram-bot==20.7
python-dotenv
